fix: 枚举失效问题 & 分离全局声明和模块化内容
